+++ hydrolib/core/dflowfm/tim/parser.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 from typing import Any, Dict, List
 
-from hydrolib.core.baseparser import BaseParser
+from hydrolib.core.base.parser import BaseParser
 
 TimData = Dict[str, List[str]]