+++ tests/test_utils.py
@@ -3,14 +3,14 @@ from pathlib import Path
 
 import pytest
 
-from hydrolib.core.dflowfm.mdu.models import Geometry, Output
-from hydrolib.core.utils import (
+from hydrolib.core.base.utils import (
     FileChecksumCalculator,
     FilePathStyleConverter,
     PathStyle,
     get_substring_between,
     str_is_empty_or_none,
 )
+from hydrolib.core.dflowfm.mdu.models import Geometry, Output
 
 from .utils import test_input_dir